+++ src/traits/i-progress/i-progress.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-extraneous-class */
+
 /*!
  * V4Fire Client Core
  * https://github.com/V4Fire/Client
@@ -14,10 +16,9 @@
 import type iBlock from 'super/i-block/i-block';
 import type { ModsDecl, ModEvent } from 'super/i-block/i-block';
 
-// eslint-disable-next-line @typescript-eslint/no-extraneous-class
 export default abstract class iProgress {
 	/**
-	 * Progress modifiers
+	 * Trait modifiers
 	 */
 	static readonly mods: ModsDecl = {
 		progress: [