+++ src/traits/i-width/i-width.ts
@@ -16,7 +16,7 @@ import type { ModsDecl } from 'super/i-block/i-block';
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class
 export default abstract class iWidth {
 	/**
-	 * Width modifiers
+	 * Trait modifiers
 	 */
 	static readonly mods: ModsDecl = {
 		width: [